+++ lib/src/mock_collection_reference.dart
@@ -3,11 +3,9 @@ import 'dart:math';
 
 import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
-import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
 import 'package:fake_cloud_firestore/src/fake_query_with_parent.dart';
 
 import 'converter.dart';
-import 'mock_collection_reference_platform.dart';
 import 'mock_document_reference.dart';
 import 'mock_query.dart';
 import 'mock_query_snapshot.dart';
@@ -26,10 +24,6 @@ class MockCollectionReference<T extends Object?> extends MockQuery<T>
 
   /// Path from the root to this collection. For example "users/USER0004/friends"
   final String _path;
-
-  // ignore: unused_field
-  final CollectionReferencePlatform _delegate =
-      MockCollectionReferencePlatform();
 
   MockCollectionReference(
     this._firestore,