+++ lib/src/mock_query_platform.dart
@@ -1,6 +1,0 @@
-import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart';
-
-class MockQueryPlatform implements QueryPlatform {
-  @override
-  dynamic noSuchMethod(Invocation invocation) => super.noSuchMethod(invocation);
-}